- change change theme position - fix responsive
