--- src/components/LoginCard/index.ts
@@ -1,3 +1,0 @@
-import LoginCard from './LoginCard.vue'
-
-export default LoginCard
